--- src/main/java/net/sphuta/tms/TmsApplication.java
@@ -1,14 +1,0 @@
-package net.sphuta.tms;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Entry point for Sphuta TMS Freelancer API.
- */
-@SpringBootApplication
-public class TmsApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(TmsApplication.class, args);
-	}
-}
