--- src/main/java/net/sphuta/tms/dto/TmsDto.java
@@ -1,5 +1,0 @@
-package net.sphuta.tms.dto;
-
-public record TmsDto(Long id, String name) {
-    // Add more fields as needed
-}
