--- src/main/java/net/sphuta/tms/exception/TmsException.java
@@ -1,8 +1,0 @@
-package net.sphuta.tms.exception;
-
-public class TmsException extends RuntimeException {
-    public TmsException(String message) {
-        super(message);
-    }
-}
-
