--- src/main/java/net/sphuta/tms/repository/TmsRepository.java
@@ -1,9 +1,0 @@
-package net.sphuta.tms.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import net.sphuta.tms.entity.TmsEntity;
-
-public interface TmsRepository extends JpaRepository<TmsEntity, Long> {
-    // Custom repository methods for TMS can be added here
-}
-
