--- src/main/java/net/sphuta/tms/service/TmsService.java
@@ -1,5 +1,0 @@
-package net.sphuta.tms.service;
-
-public interface TmsService  {
-    String getServiceStatus();
-}
