--- src/main/java/net/sphuta/tms/util/TmsUtil.java
@@ -1,6 +1,0 @@
-package net.sphuta.tms.util;
-
-public class TmsUtil {
-    // Utility methods for TMS can be added here
-}
-
