--- src/test/java/net/sphuta/tms/TmsApplicationTests.java
@@ -1,13 +1,0 @@
-package net.sphuta.tms;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TmsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
