+++ src/test/java/net/sphuta/tms/freelancer/TmsApplicationTests.java
@@ -1,0 +1,13 @@
+package net.sphuta.tms.freelancer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TmsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
